Fix rss template
This commit fixes the rss fulltext template which was originally build under the assumption of zola 0.9.0
For 0.5.1 the template has to be in a different location and can't use `escape_xml` as that was introduced in later versions.
This should do no harm because the links didn't contain xml in the first place.